🔥 :: 중복된 PasswordResetVC, SignUpVC 수정
