--- Projects/Feature/Sources/Scene/PasswordReset/Controller/PasswordResetViewController.swift
@@ -1,7 +1,0 @@
-import UIKit
-
-class PasswordResetViewController: BaseViewController {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
